pull out add_to_help_text as helper funciton
