+++ commcare-cloud/commcare_cloud/parse_help.py
@@ -1,6 +1,8 @@
+from __future__ import print_function
 from collections import namedtuple
 import os
 import subprocess
+import sys
 
 help_cache = os.path.join(os.path.dirname(__file__), 'help_cache')
 available_help_caches = {
@@ -120,3 +122,14 @@ def filtered_help_message(command, below_line, exclude_args, above_line=None):
             break
 
     return fsa.get_filtered_help_message(exclude_args)
+
+
+def add_to_help_text(parser, additional_text):
+    super_print_help = parser.print_help
+
+    def print_help(file=None):
+        if file is None:
+            file = sys.stdout
+        super_print_help(file)
+        print(additional_text, file=file)
+    parser.print_help = print_help
